world bosses bypass emoji limit

+++ index.js
@@ -546,6 +546,7 @@ bot.on('messageCreate', async message => {
       if (message.author.bot) return;
       if (message.member.roles.cache.some(role => role.id === '929941845004415049')) return;
       if (message.member.roles.cache.some(role => role.id === '932108051924783104')) return;
+      if (message.member.roles.cache.some(role => role.id === '929941845004415048')) return;
       message.delete().catch(error => {})
     }
   }
